+++ src/r_surf.c
@@ -538,7 +538,7 @@ void R_DrawSurfaceBlock16 (void)
 		pbasesource += sourcetstep;
 		lightright += lightrightstep;
 		lightleft += lightleftstep;
-		prowdest = (unsigned short *)((long)prowdest + surfrowbytes);
+		prowdest = (unsigned short *)((int32_t)prowdest + surfrowbytes);
 	}
 
 	prowdestbase = prowdest;
